requirements.txt
ansible==10.0.1
django-filter==24.2
django-ical==1.9.2
django-imagekit==5.0.0
django-webpack-loader==3.1.0
Django==5.0.6
djangorestframework==3.15.1
factory-boy==3.3.0
Faker==25.3.0
freezegun==1.5.1
graphene-django==3.2.1
graphene==3.3
httpx==0.27.0
lxml_html_clean==0.1.1
lxml==5.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
rules==3.4
stripe==9.8.0
